Move generic file preview dialog
